requirements.txt
openai>=0.27.0
requests>=2.25.0
python-dotenv>=0.19.0
streamlit>=1.28.0
pandas>=1.3.0
jinja2>=3.0.0
flask>=2.3.0
python-gitlab>=3.15.0
gitpython>=3.1.32
aiohttp>=3.8.5
plotly>=5.15.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
